Updating the UL2016 scale corrections
With this update the scale corrections for UL2016 are updated to account a few minor errors in the previous files.
The are :
UL16PreVFP --  For |Eta| between 1.5 to 2 and R9 > 0.98 the scale correction was set as 0. So effectively particles satisfying this criteria would
be removed in data (but not MC). 
UL16PostVFP --  Rhe scales for the R9>0.96 for |Eta| between 1.0 to 1.44
was set to 1

The smearings are given the same versioning as the scales for consistency but there are no changes for these.